Metadata profile suggestions from GRBio
In this document, there are mappings for existing metadata and suggested metadata that is not currently supported by fields in EML for sharing collections-level metadata. This is the work of GRBio and others, including input from the GBIF Global Strategy and Action Plan for Natural History Collections (GSAP NHC), NAP ISO 19115/INSPIRE, and the GBIF Metadata Implementation Framework Task Group.