Biodiv Registry Datafield Crosswalk - Sept 2015
From GRBio, an excel spreadsheet with crosswalk of data fields present (or not) across various biodiversity registries including: GRBio, Culture Collections, GGBN, and the Insect and Spider Collections of the World. It includes a column on NCD to evaluate what fields are (or are not) present in the NCD original draft compared to these other sites.